Add files via upload
A designer bag store page where you can add items to your cart or remove them and it displays the things in the shopping cart and gives a total of it. The billing Page takes the user information such as name, email, address, phone number, and payment methods. The checkout page shows the order summary and the total cost.